Added production tyler lower court codes
The production versions of the courts are inexplicitly different from the
stage versions. I have added a different attribute to hold this different
value. It'd be nice to only have one, but testing on stage is still useful.

None of this is document by Tyler (and they didn't answer any questions last
time I asked about this), so I anticipate this might have to stay, even though
it's hacky (and will need to be handled at the interview level).

Maybe, we could eventually save this information with the e-file proxy, closer
to where it belongs, but that'll have to be a job for later.