Stop docassemble from pre-loading tests
Some clean up I had made to tihs package earlier; if the `hypothesis` testing
library isn't on the docassemble server, docassemble will error each time it
restarts as it tries to pre-load the testing file with a class in it.

Turning this off avoids that error log (which for a while I thought was causing
issues with new logs not being made on my local DA server, but that may have
been misguided).